refactor(chrome-extension): migrate background to wxt entrypoint
